LayoutTests/fast/xpath/substring-non-positive-postion-expected.txt

Test for bug 41913 (http://bugs.webkit.org/show_bug.cgi?id=41913): XPath substring function does not correctly handle non-positive values for the position argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.evaluate("substring('abcde', 0)", document, null, XPathResult.STRING_TYPE, null).stringValue is 'abcde'
PASS document.evaluate("substring('abcde', -2)", document, null, XPathResult.STRING_TYPE, null).stringValue is 'abcde'
PASS document.evaluate("substring('abcde', 0, 5)", document, null, XPathResult.STRING_TYPE, null).stringValue is 'abcd'
PASS document.evaluate("substring('abcde', -2, 5)", document, null, XPathResult.STRING_TYPE, null).stringValue is 'ab'
PASS successfullyParsed is true

TEST COMPLETE
